requirements.txt
pytesseract~=0.3.10
Pillow~=9.5.0
aiogram~=2.25.1